Adds command to list transactional email template usage.
When an installation has many transactional emails, it is easy
to lose track of which ones are currently being used. As you can
also fall back to using templates in the source code, it is helpful
to be able to see which email templates are in use and which can be deleted.

This command allows you to see which email templates are used, and in
which scopes of the configuration that is.